+++ backend/app/api/dependencies/__init__.py
@@ -1,0 +1,1 @@
+"""API dependencies."""
