+++ backend/app/api/dependencies/agents.py
@@ -1,0 +1,12 @@
+"""Agents dependencies."""
+
+from app.core.agents import AgentsService
+
+
+def get_agent_service() -> AgentsService:
+    """Get agents service.
+
+    Returns:
+        AgentsService: Agents service.
+    """
+    return AgentsService()
