+++ backend/app/api/dependencies/companies.py
@@ -1,0 +1,12 @@
+"""Companies dependencies."""
+
+from app.core.companies import CompaniesService
+
+
+def get_companies_service() -> CompaniesService:
+    """Get companies service.
+
+    Returns:
+        CompaniesService: Companies service.
+    """
+    return CompaniesService()
