+++ backend/app/api/routes/agents.py
@@ -5,7 +5,8 @@ from typing import Annotated
 from fastapi import APIRouter
 from fastapi.params import Depends
 
-from app.api.dependencies import get_agent_service, get_current_user
+from app.api.dependencies.agents import get_agent_service
+from app.api.dependencies.auth import get_current_user
 from app.api.exceptions.agents import AgentNotFound
 from app.api.schemes.agents import AgentResponse
 from app.core.agents import (
