+++ backend/app/api/routes/offers.py
@@ -6,11 +6,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends
 from fastapi.responses import StreamingResponse
 
-from app.api.dependencies import (
-    get_admin,
-    get_current_user,
-    get_offers_service,
-)
+from app.api.dependencies.auth import get_admin, get_current_user
+from app.api.dependencies.offers import get_offers_service
 from app.api.exceptions.offers import BadOfferFile, OfferNotFound
 from app.api.schemes.offers import (
     OfferCreate,
