+++ backend/app/api/routes/works.py
@@ -7,7 +7,7 @@ from typing import Annotated
 
 from fastapi import APIRouter, Depends
 
-from app.api.dependencies import get_admin, get_current_user
+from app.api.dependencies.auth import get_admin, get_current_user
 from app.api.exceptions.works import WorkNotFound
 from app.api.schemes.works import (
     WorkCreate,
